+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -282,6 +282,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
     }
 
     @Test
+    @TestMetadata("UnusedGraphInputsInGraphExtensionAreReported.kt")
+    public void testUnusedGraphInputsInGraphExtensionAreReported() {
+      runTest("compiler-tests/src/test/data/diagnostic/dependencygraph/UnusedGraphInputsInGraphExtensionAreReported.kt");
+    }
+
+    @Test
     @TestMetadata("UsefulHintsForMissingImplDeps.kt")
     public void testUsefulHintsForMissingImplDeps() {
       runTest("compiler-tests/src/test/data/diagnostic/dependencygraph/UsefulHintsForMissingImplDeps.kt");